+++ src/components/forms/FooterNewInvoice.js
@@ -2,8 +2,11 @@ import { useFormikContext } from 'formik';
 import { useDispatch } from "react-redux";
 import { addNewInvoice } from '../../store/invoicesActions';
 import { closeNewFormModal } from "../../store/newFormModalStatusSlice";
+import { createInvoiceId } from '../../utils/consts';
 import { Button } from '../../UI/Button';
 import { getItemTotal, getPaymentDue, getTotal } from '../../utils/consts';
+
+
 
 export const FooterNewInvoice = () => {
 
@@ -15,11 +18,11 @@ export const FooterNewInvoice = () => {
     }
 
     const saveAsDraftHandler = () => {
-        console.log(values.id);
+        values.id = createInvoiceId();
         values.status = 'draft';
         values.paymentDue = getPaymentDue(values.createdAt, values.paymentTerms);
         if (values.items.length > 0) {
-            values.items.map((item) => item.total === getItemTotal(item))
+            values.items.map((item) => item.total = getItemTotal(item))
 
             values.total = getTotal(values.items);
         }
